% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{optim_torch}
\alias{optim_torch}
\title{A \code{stats::optim} like optimizer based on torch}
\usage{
optim_torch(params, fn, method, control = list(), ...)
}
\arguments{
\item{params}{parameters to \code{fn} as a tensor}

\item{fn}{a function taking the \code{params} as input}

\item{method}{the method to be used for optimization as a string}

\item{control}{a control list with two slots, \code{maxiter} and \code{abstol}}

\item{...}{Remaining parameters are passed to the selected method}
}
\value{
A list
}
\description{
NOTE: Experimental
}
\examples{
library(torch)
library(torchoptim)
# from the R docs of stats::optim
fr <- function(x) {
  x1 <- x[1]
  x2 <- x[2]
  100 * (x2 - x1 * x1)^2 + (1 - x1)^2
}
optim_torch(
  torch_tensor(c(-1.2, 1), requires_grad = TRUE),
  fr,
  method = "lbfgs",
  control = list(maxiter = 10)
)
}
